# app/swagger.yaml
...
swagger: "2.0"
info:
  title: "Banking API"
  description: "API documentation for the Simple Banking application (CodeJam Competition)."
  version: "1.1.0" # Updated version
host: "localhost:5000" # Adjust if deployed elsewhere
basePath: "/api"
schemes:
  - "http" # Change to https if using HTTPS
  # - "https"
...
securityDefinitions:
  BearerAuth: # Renamed for clarity
    type: apiKey
    name: Authorization
    in: header
    description: 'JWT Authorization header using the Bearer scheme. Example: "Authorization: Bearer {token}"'

# Apply BearerAuth security globally to endpoints that require it,
# or specify per-operation using 'security:' keyword.
# security:
#   - BearerAuth: []
paths:
...
  /auth/register:
    post:
      summary: "Register New User"
      description: "Registers a new user account."
      tags: ["Authentication"]
      consumes: ["application/json"]
      produces: ["application/json"]
      parameters:
        - in: "body"
          name: "body"
          required: true
          description: "User registration details."
          schema:
            type: "object"
            required: [email, password]
            properties:
              email:
                {
                  type: "string",
                  format: "email",
                  example: "testuser@example.com",
                }
              password:
                {
                  type: "string",
                  format: "password",
                  example: "Str0ngP@ssw0rd!",
                }
              username:
                {
                  type: "string",
                  example: "testuser123",
                  description: "Optional. Generated if not provided.",
                }
              first_name: { type: "string", example: "Test" }
              last_name: { type: "string", example: "User" }
      responses:
        201:
          description: "User registered successfully."
          schema:
            $ref: "#/definitions/UserResponse"
        400: { $ref: "#/definitions/ErrorResponse" }
        409: { $ref: "#/definitions/ErrorResponse" } # Conflict (username/email exists)
        500: { $ref: "#/definitions/ErrorResponse" }
  /auth/login:
    post:
      summary: "User Login"
      description: "Authenticates a user with email/username and password, returns JWT tokens."
      tags: ["Authentication"]
      consumes: ["application/json"]
      produces: ["application/json"]
      parameters:
        - in: "body"
          name: "body"
          required: true
          description: "User login credentials."
          schema:
            type: "object"
            required: [password]
            properties:
              email:
                {
                  type: "string",
                  format: "email",
                  example: "testuser@example.com",
                  description: "Login with email OR username.",
                }
              username:
                {
                  type: "string",
                  example: "testuser123",
                  description: "Login with email OR username.",
                }
              password:
                {
                  type: "string",
                  format: "password",
                  example: "Str0ngP@ssw0rd!",
                }
      responses:
        200:
          description: "Login successful."
          schema:
            type: "object"
            properties:
              message: { type: "string", example: "Login successful" }
              access_token:
                { type: "string", description: "JWT Access Token (fresh)" }
              refresh_token:
                { type: "string", description: "JWT Refresh Token" }
              user: { $ref: "#/definitions/User" }
        400: { $ref: "#/definitions/ErrorResponse" } # Missing fields
        401: { $ref: "#/definitions/ErrorResponse" } # Invalid credentials
        403: { $ref: "#/definitions/ErrorResponse" } # Inactive user
  /auth/refresh:
    post:
      summary: "Refresh Access Token"
      description: "Obtains a new (non-fresh) access token using a valid refresh token."
      tags: ["Authentication"]
      produces: ["application/json"]
      security: # Requires refresh token
        - BearerAuth: []
      responses:
        200:
          description: "Access token refreshed successfully."
          schema:
            type: "object"
            properties:
              access_token:
                {
                  type: "string",
                  description: "New JWT Access Token (non-fresh)",
                }
        401: { $ref: "#/definitions/ErrorResponse" } # Invalid/expired refresh token, or inactive user
  /auth/logout:
    post:
      summary: "User Logout"
      description: "Logs out the user by adding the current token's JTI to the server-side blocklist."
      tags: ["Authentication"]
      produces: ["application/json"]
      security: # Requires valid access or refresh token
        - BearerAuth: []
      responses:
        200:
          description: "Successfully logged out."
          schema:
            type: "object"
            properties:
              message:
                {
                  type: "string",
                  example: "Successfully logged out. Access token revoked.",
                }
        400: { $ref: "#/definitions/ErrorResponse" } # e.g., Missing JTI
        401: { $ref: "#/definitions/ErrorResponse" } # Invalid/expired token
  /auth/verify:
    post:
      summary: "Verify Token"
      description: "Checks if the provided JWT token is valid, not expired, not revoked, and belongs to an active user."
      tags: ["Authentication"]
      produces: ["application/json"]
      security: # Requires valid access token
        - BearerAuth: []
      responses:
        200:
          description: "Token is valid and active."
          schema:
            type: "object"
            properties:
              message: { type: "string", example: "Token is valid and active" }
              verified: { type: "boolean", example: true }
              user_id: { type: "integer", example: 1 }
              role: { type: "string", example: "user" }
        401: { $ref: "#/definitions/ErrorResponse" } # Invalid/expired/revoked token
        403: { $ref: "#/definitions/ErrorResponse" } # User inactive
        404: { $ref: "#/definitions/ErrorResponse" } # User not found
  /auth/profile:
    get:
      summary: "Get User Profile"
      description: "Retrieves the profile information of the currently authenticated user."
      tags: ["Authentication"]
      produces: ["application/json"]
      security: # Requires valid access token
        - BearerAuth: []
      responses:
        200:
          description: "User profile retrieved successfully."
          schema:
            $ref: "#/definitions/User" # Return the user object directly
        401: { $ref: "#/definitions/ErrorResponse" } # Invalid/expired token
        403: { $ref: "#/definitions/ErrorResponse" } # User inactive
        404: { $ref: "#/definitions/ErrorResponse" } # User not found
  /auth/change-password:
    post:
      summary: "Change Password"
      description: "Allows the authenticated user to change their password. Requires a fresh token (from login)."
      tags: ["Authentication"]
      consumes: ["application/json"]
      produces: ["application/json"]
      security: # Requires fresh access token
        - BearerAuth: []
      parameters:
        - in: "body"
          name: "body"
          required: true
          description: "Password change details."
          schema:
            type: "object"
            required: [current_password, new_password]
            properties:
              current_password:
                {
                  type: "string",
                  format: "password",
                  example: "Str0ngP@ssw0rd!",
                }
              new_password:
                {
                  type: "string",
                  format: "password",
                  example: "NewerP@ssw0rd123!",
                }
      responses:
        200:
          description: "Password changed successfully."
          schema:
            type: "object"
            properties:
              message:
                { type: "string", example: "Password changed successfully" }
        400: { $ref: "#/definitions/ErrorResponse" } # Validation errors (complexity, same password)
        401: { $ref: "#/definitions/ErrorResponse" } # Incorrect current password or non-fresh token
        500: { $ref: "#/definitions/ErrorResponse" }

  # --- Admin ---
  /auth/users:
    get:
      summary: "Get All Users (Admin)"
      description: "Retrieves a paginated list of all users. Requires admin privileges."
      tags: ["Admin"]
      produces: ["application/json"]
      security: # Requires admin role
        - BearerAuth: []
      parameters:
        - $ref: "#/parameters/PageParam"
        - $ref: "#/parameters/PerPageParam"
      responses:
        200:
          description: "List of users retrieved successfully."
          schema:
            type: "object"
            properties:
              users:
                type: "array"
                items: { $ref: "#/definitions/User" }
              page: { type: "integer", example: 1 }
              per_page: { type: "integer", example: 10 }
              total_users: { type: "integer", example: 50 }
              total_pages: { type: "integer", example: 5 }
        400: { $ref: "#/definitions/ErrorResponse" } # Invalid pagination
        401: { $ref: "#/definitions/ErrorResponse" } # Invalid/expired token
        403: { $ref: "#/definitions/ErrorResponse" } # Admin required
  /auth/user/{user_id}:
    delete:
      summary: "Delete User (Admin)"
      description: "Deletes a user by their ID (Hard Delete). Requires admin privileges."
      tags: ["Admin"]
      produces: ["application/json"]
      security: # Requires admin role
        - BearerAuth: []
      parameters:
        - name: "user_id"
          in: "path"
          required: true
          description: "ID of the user to delete."
          type: "integer"
      responses:
        200:
          description: "User deleted successfully."
          schema:
            type: "object"
            properties:
              message:
                {
                  type: "string",
                  example: "User ID 5 (someuser) deleted successfully",
                }
        401: { $ref: "#/definitions/ErrorResponse" } # Invalid/expired token
        403: { $ref: "#/definitions/ErrorResponse" } # Admin required or trying to delete self
        404: { $ref: "#/definitions/ErrorResponse" } # User not found
        500: { $ref: "#/definitions/ErrorResponse" }

  # --- Accounts ---
...
parameters:
  PageParam:
    name: "page"
    in: "query"
    description: "Page number for pagination (default: 1)."
    required: false
    type: "integer"
    default: 1
    minimum: 1
  PerPageParam:
    name: "per_page"
    in: "query"
    description: "Number of items per page (default: 10 or 20, max: 100)."
    required: false
    type: "integer"
    default: 10 # Or 20 depending on endpoint default
    minimum: 1
    maximum: 100

# --- Definitions (Reusable Schemas) ---
definitions:
  ErrorResponse:
    type: "object"
    properties:
      error:
        type: "string"
        example: "Detailed error message"
  User:
    type: "object"
    properties:
      id: { type: "integer", example: 1 }
      username: { type: "string", example: "testuser123" }
      email:
        { type: "string", format: "email", example: "testuser@example.com" }
      first_name: { type: "string", example: "Test", nullable: true }
      last_name: { type: "string", example: "User", nullable: true }
      created_at:
        { type: "string", format: "date-time", example: "2023-10-27T10:00:00Z" }
      role: { type: "string", example: "user" }
      is_active: { type: "boolean", example: true }
  UserResponse: # For registration response
    type: "object"
    properties:
      message: { type: "string", example: "User registered successfully" }
      user: { $ref: "#/definitions/User" }
...